Change catches on express routes to res.error(err) after console.log
