+++ express/routes/jobsRoutes.js
@@ -20,6 +20,7 @@ module.exports = (knex) => {
       })
       .catch(err => {
         console.log(`Error: could not POST to /jobs ${err}`);
+        res.error(err);
       });
     });
 
@@ -39,6 +40,7 @@ module.exports = (knex) => {
       })
       .catch(err => {
         console.log(`Error: could not POST to /jobs/${req.params.id} ${err}`)
+        res.error(err);
       });
     });
